change the fragmentTransition back because I cannot have getSupportManager working
